+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- The `Authenticate` method now parses the authentication response and writes it
+  to the token file, without the need to call `ParseAuthenticationResponse`.
+  This change breaks the API.
+  [cyberark/conjur-authn-k8s-client#180](https://github.com/cyberark/conjur-authn-k8s-client/issues/180)
 
 ## [0.19.0] - 2020-10-08
 ### Added